:recycle: Refactor Entity to use BigInteger as ID type

+++ src/main/java/br/com/metaway/petshop/models/Entity.java
@@ -1,5 +1,7 @@
 package br.com.metaway.petshop.models;
 
+import java.math.BigInteger;
+
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -15,13 +17,13 @@ import jakarta.persistence.MappedSuperclass;
 public class Entity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	protected Integer id;
+	protected BigInteger id;
 
-	public Integer getId() {
+	public BigInteger getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(BigInteger id) {
 		this.id = id;
 	}
 }
